refactor(ORCH-047): rehome room_types dataclasses into repository files and delete room_types.py
