+++ tests/test_sse_persistence.py
@@ -12,7 +12,7 @@ from jellyswipe.db_runtime import (
 )
 from jellyswipe.db_uow import DatabaseUnitOfWork
 from jellyswipe.migrations import build_sqlite_url, upgrade_to_head
-from jellyswipe.room_types import RoomStatusSnapshot
+from jellyswipe.repositories.rooms import RoomStatusSnapshot
 
 
 @pytest.fixture(autouse=True)
